Thruster: Scaling HUD Bar
Implemented the scaling HUD bar feature by creating a cooling system that allows the player to boost their thrusters for 2 seconds and then needing to charge them for 4 seconds for a full charge. The player can use their thrusters as long as the charge is above 0%. Created a UI box scaling on the x axis that represents this cool down and charge up process.